v0.0.14 - relocate scalar option functions
