+++ src/graphql/resolver/create_resolver_function/get_resolver_input/create_internal_input/get_parent_value/mod.rs
@@ -5,9 +5,8 @@ use log::debug;
 use crate::{
     configuration::subgraph::entities::service_entity_field::ServiceEntityFieldConfig,
     data_sources::sql::services::ResponseRow, graphql::resolver::ServiceResolver,
+    scalar_option::ScalarOption,
 };
-
-mod response_row_to_document;
 
 impl ServiceResolver {
     /// Gets the parent value from the context, which is either a document or a response row.
@@ -38,21 +37,21 @@ impl ServiceResolver {
                         // Map the value into a Document, which is what the resolver expects.
                         Some(response_row) => match response_row {
                             ResponseRow::SqLite(rr) => {
-                                Some(ServiceResolver::sqlite_response_row_to_doc(
+                                Some(ScalarOption::sqlite_response_row_to_input_doc(
                                     rr,
                                     as_type_field,
                                     field_name,
                                 )?)
                             }
                             ResponseRow::MySql(rr) => {
-                                Some(ServiceResolver::mysql_response_row_to_doc(
+                                Some(ScalarOption::mysql_response_row_to_input_doc(
                                     rr,
                                     as_type_field,
                                     field_name,
                                 )?)
                             }
                             ResponseRow::Postgres(rr) => {
-                                Some(ServiceResolver::postgres_response_row_to_doc(
+                                Some(ScalarOption::postgres_response_row_to_input_doc(
                                     rr,
                                     as_type_field,
                                     field_name,
