+++ src/scalar_option/mod.rs
@@ -1,5 +1,8 @@
 use serde::{Deserialize, Serialize};
 
+pub mod mysql_response_row_to_input_doc;
+pub mod postgres_response_row_to_input_doc;
+pub mod sqlite_response_row_to_input_doc;
 pub mod to_async_graphql_value;
 pub mod to_bson_element_type;
 pub mod to_evalexpr;
